Add support to process MCAP files
